+++ lib/semantic_logger.rb
@@ -15,6 +15,7 @@ module SemanticLogger
   autoload :Loggable,           "semantic_logger/loggable"
   autoload :Processor,          "semantic_logger/processor"
   autoload :Subscriber,         "semantic_logger/subscriber"
+  autoload :SyncProcessor,      "semantic_logger/sync_processor"
   autoload :Utils,              "semantic_logger/utils"
 
   module Concerns